go into lab
