+++ docs/src/pages/components/autocomplete/Playground.tsx
@@ -1,5 +1,4 @@
 /* eslint-disable no-use-before-define */
-
 import React from 'react';
 import Autocomplete from '@material-ui/lab/Autocomplete';
