+++ packages/material-ui-lab/src/index.js
@@ -6,5 +6,6 @@ export { default as SpeedDialAction } from './SpeedDialAction';
 export { default as SpeedDialIcon } from './SpeedDialIcon';
 export { default as ToggleButton } from './ToggleButton';
 export { default as ToggleButtonGroup } from './ToggleButtonGroup';
-export { default as TreeView } from './TreeView';
 export { default as TreeItem } from './TreeItem';
+export { default as TreeView } from './TreeView';
+export { default as useAutocomplete, createFilterOptions } from './useAutocomplete';
