+++ packages/material-ui-lab/src/useAutocomplete/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './useAutocomplete';
+export * from './useAutocomplete';
