+++ packages/material-ui-lab/src/useAutocomplete/index.js
@@ -1,0 +1,1 @@
+export { default, createFilterOptions } from './useAutocomplete';
